pull code into controller, rename files
