--- service/routes/index.js
@@ -1,19 +1,0 @@
-"use strict"
-
-/*
-
-  Index.js is the default export when a folder is required.
-  Hence, when we require /routes, we are requiring this index.js file
-  which hence should include all the routes.
-
-*/
-
-// Adding dependent routes
-const lifeCycleRoutes = require('./lifeCycleRoutes')
-
-// Packaging and exporting
-module.exports = function(app, db) {
-  lifeCycleRoutes(app, db)
-
-
-}
